Rename fragment A and fragment B
