+++ app/src/main/java/shevtsov/daniil/incrementalreader/creation/CreationFragment.kt
@@ -1,11 +1,11 @@
-package shevtsov.daniil.incrementalreader.fragmentb
+package shevtsov.daniil.incrementalreader.creation
 
 import androidx.fragment.app.Fragment
 import shevtsov.daniil.incrementalreader.R
 import shevtsov.daniil.incrementalreader.core.util.viewLifecycleLazy
-import shevtsov.daniil.incrementalreader.databinding.FragmentBBinding.bind
+import shevtsov.daniil.incrementalreader.databinding.FragmentCreationBinding.bind
 
-class FragmentB : Fragment(R.layout.fragment_b) {
+class CreationFragment : Fragment(R.layout.fragment_creation) {
 
     private val binding by viewLifecycleLazy { bind(requireView()) }
