+++ app/src/main/java/shevtsov/daniil/incrementalreader/creation/CreationFragmentViewModel.kt
@@ -1,0 +1,10 @@
+package shevtsov.daniil.incrementalreader.creation
+
+import androidx.lifecycle.ViewModel
+import javax.inject.Inject
+
+class CreationFragmentViewModel @Inject constructor(
+
+) : ViewModel() {
+
+}
