--- app/src/main/java/shevtsov/daniil/incrementalreader/fragmenta/FragmentAViewModel.kt
@@ -1,9 +1,0 @@
-package shevtsov.daniil.incrementalreader.fragmenta
-
-import androidx.lifecycle.ViewModel
-import javax.inject.Inject
-
-class FragmentAViewModel @Inject constructor(
-
-) : ViewModel() {
-}
