--- app/src/main/java/shevtsov/daniil/incrementalreader/fragmentb/FragmentBViewModel.kt
@@ -1,10 +1,0 @@
-package shevtsov.daniil.incrementalreader.fragmentb
-
-import androidx.lifecycle.ViewModel
-import javax.inject.Inject
-
-class FragmentBViewModel @Inject constructor(
-
-) : ViewModel() {
-
-}
