+++ app/src/main/java/shevtsov/daniil/incrementalreader/main/MainFragmentViewModel.kt
@@ -1,0 +1,9 @@
+package shevtsov.daniil.incrementalreader.main
+
+import androidx.lifecycle.ViewModel
+import javax.inject.Inject
+
+class MainFragmentViewModel @Inject constructor(
+
+) : ViewModel() {
+}
