+++ app/src/main/res/navigation/main_navigation.xml
@@ -3,20 +3,20 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main_navigation"
-    app:startDestination="@id/fragment_a">
+    app:startDestination="@id/main_fragment">
 
     <fragment
-        android:id="@+id/fragment_a"
-        android:name="shevtsov.daniil.incrementalreader.fragmenta.FragmentA"
-        tools:layout="@layout/fragment_a">
+        android:id="@+id/main_fragment"
+        android:name="shevtsov.daniil.incrementalreader.main.MainFragment"
+        tools:layout="@layout/fragment_main">
         <action
-            android:id="@+id/navigate_to_b_direction"
-            app:destination="@id/fragment_b" />
+            android:id="@+id/navigate_to_creation_direction"
+            app:destination="@id/creation_fragment" />
     </fragment>
 
     <fragment
-        android:id="@+id/fragment_b"
-        android:name="shevtsov.daniil.incrementalreader.fragmentb.FragmentB"
-        tools:layout="@layout/fragment_b" />
+        android:id="@+id/creation_fragment"
+        android:name="shevtsov.daniil.incrementalreader.creation.CreationFragment"
+        tools:layout="@layout/fragment_creation" />
 
 </navigation>
